Fix public space tab rename optimism and revert duplicate tab guard
